refactor XmlValidateConst separate module
